% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_ds3100.R
\name{install_ds3100}
\alias{install_ds3100}
\title{Install DS3100 Packages}
\usage{
install_ds3100()
}
\value{
Returns messages and a numeric value:

\item{0}{Successful package installations}

\item{1}{Failed package installations}
}
\description{
A single function to install all R packages for
the DS3100 course
}
\examples{
\donttest{
install_ds3100()}

}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
