% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tinytex.R
\name{install_tinytex}
\alias{install_tinytex}
\title{Install \code{\link{tinytex}}}
\usage{
install_tinytex()
}
\value{
Returns messages and a numeric value:

\item{0}{Successful \code{\link{tinytex}} installation}

\item{1}{Failed \code{\link{tinytex}} installation}
}
\description{
A single function to install \code{\link{tinytex}}
}
\examples{
\donttest{
install_tinytex()}

}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
