% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolchainR.R
\name{toolchainR}
\alias{toolchainR}
\title{Install R Toolchains}
\usage{
toolchainR()
}
\value{
Returns messages and a numeric value:

\item{0}{Successful toolchain installation}

\item{1}{Failed toolchain installation}
}
\description{
A single function to install R toolchains based on your
R version and operating system. Follow the instructions
}
\examples{
if(interactive()){
  toolchainR()}

}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com>
}
